requirements.txt
# Core image + config stack
numpy>=1.26.0
opencv-python>=4.8.0
PyYAML>=6.0.1
pytesseract>=0.3.10
pandas>=2.0.0

# OCR backends
# For CUDA installs, replace `paddlepaddle` below with the matching
# `paddlepaddle-gpu` wheel for your OS/CUDA runtime.
paddlepaddle>=3.3.0
paddleocr[doc-parser]>=3.4.0,<3.5
transformers>=4.52.0,<5

# Optional dataset / script support
mistralai>=1.12.4

# Optional GLM fallback (vendored source)
-e ./vendor/GLM-OCR
